+++ src/extra/recipes/__init__.py
@@ -1,4 +1,7 @@
 
 from .grating import Grating2DCalibration # noqa
 from .cookiebox import CookieboxCalibration # noqa
+from .grating import Grating1DCalibration # noqa
+from .cookiebox_deconvolve import TOFResponse # noqa
+from .vs import VSLight # noqa